Alpha version 0.1.3.2! User&save Update completed!
Fully working user system Users are now objects that now refers to files by the name of the user a player and map object is loaded from a text file with the use serialization. Now user can save their game and login and logout. the game also auto saves when quiting the game or logging out. major day for adventure!